+++ pdf/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improve TTF Writer compatibility
 - Apply THE BIDIRECTIONAL ALGORITHM using dart_bidi [Milad akarie]
+- Add Inseparable Widget
 
 ## 3.8.4